# Performance optimization on recommendation compute