--- a/lux/executor/PandasExecutor.py
+++ b/lux/executor/PandasExecutor.py
@@ -456,13 +456,9 @@ class PandasExecutor(Executor):
                 warn_msg += f"\tdf['{attr}'] = pd.to_datetime(df['{attr}'], format='<replace-with-datetime-format>')\n"
             warn_msg += "\nSee more at: https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.to_datetime.html"
             warnings.warn(warn_msg, stacklevel=2)
-
-    def _is_datetime_string(self, series):
-        if len(series) > 100:
-            series = series.sample(100)
-
+    @staticmethod
+    def _is_datetime_string(series):
         if series.dtype == object:
-
             not_numeric = False
             try:
                 pd.to_numeric(series)