Rewrote the following code which was slowing down `t_rec`: 
- remove sampling in `is_datetime_string`, since sampling was slower than no sampling (425ms --> 480ms)
- changed `scipy.stats.spearmanr` to `scipy.stats.pearsonr`(360ms --> 2.2ms, huge performance gain!)

Overall, with these changes the 1st print time for 1 million datapoints on Airbnb (12 cols, no sampling) went from around 6.5sec down to around 3.8, so around 1.8x faster.